Migrate script.py into modular files
